+++ api/tags/delete.ts
@@ -10,7 +10,8 @@ import { initDatabase } from '../../db/db';
 import { Tag } from '../../db/models/tag';
 import { AuthMiddleware } from '../../utils/auth-middleware';
 import { getUserEmail } from '../../utils/get-user-email';
-import { verifyCollection } from '../../utils/verify-collection';
+import { getCollectionUser } from '../../utils/get-collection-user';
+import createHttpError from 'http-errors';
 
 initDatabase();
 
@@ -20,7 +21,11 @@ const deleteHandler: APIGatewayProxyHandler = async (event, _context) => {
   try {
     const collectionId = event.pathParameters.collectionId;
     const userEmail = getUserEmail(event);
-    await verifyCollection(collectionId, userEmail);
+    const collectionUser = await getCollectionUser(collectionId, userEmail);
+
+    if (collectionUser.role !== 'Admin' && collectionUser.role !== 'Standard') {
+      throw createHttpError(401, `User Role doesn't allow tag deletion`);
+    }
 
     const item = await Tag.findOneAndDelete(
       {
