+++ api/users/delete.ts
@@ -4,12 +4,13 @@ import middy from '@middy/core';
 import cors from '@middy/http-cors';
 import httpErrorHandler from '@middy/http-error-handler';
 import { APIGatewayProxyHandler } from 'aws-lambda';
+import createHttpError from 'http-errors';
 
 import { initDatabase } from '../../db/db';
 import { CollectionUser } from '../../db/models/collection-user';
 import { AuthMiddleware } from '../../utils/auth-middleware';
+import { getCollectionUser } from '../../utils/get-collection-user';
 import { getUserEmail } from '../../utils/get-user-email';
-import { verifyCollection } from '../../utils/verify-collection';
 
 initDatabase();
 
@@ -21,7 +22,14 @@ const removeUserHandler: APIGatewayProxyHandler = async (event, _context) => {
   try {
     const requestUserEmail = getUserEmail(event);
 
-    await verifyCollection(collectionId, requestUserEmail);
+    const collectionUser = await getCollectionUser(
+      collectionId,
+      requestUserEmail,
+    );
+
+    if (collectionUser.role !== 'Admin') {
+      throw createHttpError(401, `User Role doesn't allow user deletion`);
+    }
 
     const item = await CollectionUser.findByIdAndDelete(id);
